Headers: split mappings.h into focused subset headers
mappings.h was doing five jobs in one file: GL include selection,
threading primitives, logging macros, OS detection / backend defines,
and miscellaneous alpha constants. Hard to evolve, even harder to
include selectively.

Split it into three focused headers, with mappings.h kept as a thin
aggregator so external callers that `#include "mappings.h"` keep
working:

  - GLBackend.h    — desktop GL / GLES2 / iOS / Qt branch selection,
                     INIT_GL, GL_EXT_FUNC, checkGlError, ALPHA_*.
  - Threading.h    — typedef std::mutex LOCK + LOCK_* macros.
  - LogMacros.h    — LOGI / LOGE / LOGITIMETOPIC dispatch (Android
                     logcat vs printf), LOG_TAG.

mappings.h now just declares MAX_SCENES + STORE_RELEASE, then
includes the three pieces. Drops the accumulated dead ifdef
branches (TARGET_LOGGER_* commented-out block stays but is the only
relic).

Internal gainlib code can start preferring the focused headers when
adding new files; existing files continue to compile unchanged.

Verified: cmake --build clean, downstream TriRamp self-test green,
GL client opens.
